#1117 Implement HeadOnce/BodyOnce caching and reuse them in RqOnce
HeadOnce now caches the head iterable on first access via Sticky, so
subsequent head() calls return the same data while body() stays a
pass-through. BodyOnce caches the body content as a String and emits a
fresh InputStream over it on every body() call, leaving head()
unwrapped. RqOnce composes the two — RqOnce(req) is now equivalent to
new HeadOnce(new BodyOnce(req)) — and the resolved puzzle comment is
removed.
