+++ src/main/java/org/takes/rq/HeadOnce.java
@@ -5,13 +5,20 @@
 package org.takes.rq;
 
 import lombok.EqualsAndHashCode;
+import org.cactoos.scalar.IoChecked;
+import org.cactoos.scalar.Sticky;
 import org.takes.Request;
 
 /**
  * Request decorator that prevents multiple reads of head content.
  *
- * <p>Skeleton without caching — to be implemented next; the existing test
- * {@link HeadOnceTest#cachesHeadOnFirstAccess} fails until caching is added.
+ * <p>This decorator caches the request headers on first access, ensuring that
+ * subsequent calls to {@code head()} return the same cached data. The body
+ * is delegated to the original request without caching. This is useful when
+ * the head is computed lazily and we want to make sure it is not produced
+ * more than once, while still allowing the body to be streamed normally.
+ *
+ * <p>The class is immutable and thread-safe.
  *
  * @since 2.0
  */
@@ -23,6 +30,11 @@ public final class HeadOnce extends RqWrap {
      * @param req Original request
      */
     public HeadOnce(final Request req) {
-        super(req);
+        super(
+            new RequestOf(
+                new IoChecked<>(new Sticky<>(req::head))::value,
+                req::body
+            )
+        );
     }
 }
